Fix command and property pad routing issue
Disabled routing of commands through the active document
when the document doesn't have the focus. This was added to
avoid the property pad to be reset when the focus was temporarily
moved away from the designer, but this change is caused other
problems, such as for example editor commands being executed
when the focus is in a pad.

Instead of permanently routing commands to the active document,
the solution is to do it only when looking for a property pad provider.

This fixes the following bugs:
7033 - Select All does not work in the unified search box
7151 - Wrong keyboard focus